use a custom Time type for parsing TriMet times
Times returned from the TriMet API are not in the proper RFC3339 format, as
they lack the "Z" specifier separating (nano)seconds from timezone.
Additionally, there are parsing bugs in Go that lead to an inability to parse a
proper RFC3339 string that includes a timezone.

This avoids those issues by stripping all TZ info from a datetime and
forcing the PST location after successfully parsing the rest.